Hoàn thành phần register (thử nghiệm)
