--- css/register-LeDinhHuy.css
@@ -1,176 +1,0 @@
-* {
-  padding: 0;
-  margin: 0;
-  box-sizing: border-box;
-}
-
-body {
-  height: 1000px;
-}
-
-.register_header {
-  background-color: black;
-}
-.register_header .container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-.register_header .container .logo {
-  color: rgb(225, 47, 225);
-  font-size: 50px;
-  display: inline-block;
-  text-decoration: none;
-  font-family: "Tilt Prism";
-}
-.register_header .container .menu .shop {
-  color: rgb(201, 201, 201);
-  font-size: 30px;
-}
-.register_header .container .menu a {
-  position: relative;
-  display: inline-block;
-  color: whitesmoke;
-  text-decoration: none;
-  font-size: 20px;
-}
-.register_header .container .menu a::after {
-  position: absolute;
-  width: 0%;
-  height: 2px;
-  content: "";
-  bottom: -3px;
-  left: 0;
-  background-color: rgb(225, 47, 225);
-  transform: scale(0);
-}
-.register_header .container .menu a:hover {
-  color: rgb(225, 47, 225);
-  transition: all 0.5s;
-}
-.register_header .container .menu a:hover::after {
-  width: 100%;
-  transform: scaleX(1);
-  transition: 0.5s transform;
-}
-
-.register_menu {
-  background-color: rgb(201, 201, 201);
-}
-.register_menu .container {
-  margin-left: 30px;
-}
-.register_menu .container .home {
-  color: rgb(225, 47, 225);
-  font-weight: 500;
-}
-.register_menu .container .home:hover {
-  color: red;
-  transition: all 1s;
-}
-.register_menu .container a {
-  display: inline-block;
-  text-decoration: none;
-  font-size: 20px;
-  color: black;
-  padding: 0px 20px;
-  margin: 10px 0px;
-}
-.register_menu .container a:hover {
-  color: rgb(225, 47, 225);
-  transition: all 0.5s linear;
-}
-
-.register_form {
-  background-color: white;
-}
-.register_form h1 {
-  font-size: 50px;
-}
-.register_form .container .frm_register {
-  background-color: white;
-  border: none;
-}
-.register_form .container .frm_register .register__form .form__input .component__input input {
-  font-size: 30px;
-  color: black;
-}
-.register_form .container .frm_register .register__form .form__input .component__input .text_effect {
-  position: relative;
-  font-family: sans-serif;
-  text-transform: uppercase;
-  font-size: 30px;
-  letter-spacing: 4px;
-  overflow: hidden;
-  background: linear-gradient(60deg, rgb(225, 47, 225), white, red);
-  background-repeat: no-repeat;
-  background-size: 80%;
-  animation: animate 3s linear infinite;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: rgba(255, 255, 255, 0);
-  animation: text_effect linear infinite 10s;
-}
-.register_form .container .frm_register .register__form .form__input .component__input .gender:hover {
-  color: rgb(225, 47, 225);
-  transition: all 1s;
-}
-.register_form .coolBeans {
-  border: 2px solid red;
-  border-radius: 30px;
-  width: 200px;
-  height: 60px;
-  color: red;
-  font-family: roboto;
-  font-size: 30px;
-  font-weight: 400;
-  overflow: hidden;
-  position: relative;
-  text-decoration: none;
-  transition: 0.2s transform ease-in-out;
-  will-change: transform;
-  z-index: 0;
-  background-color: white;
-}
-.register_form .coolBeans::after {
-  background-color: rgb(225, 47, 225);
-  border-radius: 0.5rem;
-  content: "";
-  display: block;
-  height: 100%;
-  width: 100%;
-  position: absolute;
-  left: 0;
-  top: 0;
-  transform: translate(-100%, 0) rotate(10deg);
-  transform-origin: top left;
-  transition: 0.3s transform ease-out;
-  will-change: transform;
-  z-index: -1;
-}
-.register_form .coolBeans:hover::after {
-  transform: translate(0, 0);
-}
-.register_form .coolBeans:hover {
-  border: 2px solid transparent;
-  color: whitesmoke;
-  transform: scale(1);
-  will-change: transform;
-}
-
-@keyframes text_effect {
-  0% {
-    background-position: -500%;
-  }
-  100% {
-    background-position: 500%;
-  }
-}
-.register_footer .container .footer_register_info .footer_register_title {
-  font-size: 30px;
-}
-.register_footer .container .footer_register_info p {
-  font-size: 20px;
-}
-.register_footer .container .footer_register_info .register_more_info {
-  font-size: 20px;
-}
