--- models/register_OOP-LeDinhHuy.js
@@ -1,9 +1,0 @@
-//Class sign in of user (save user's info)
-export class CustomerInfo {
-    email = ''
-    password = ''
-    name = ''
-    gender = ''
-    phone = ''
-    passwordConfirm = ''
-}
